fastapi_ai_backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
python-jose
passlib[bcrypt]
openai
psycopg2-binary
sqlalchemy
fastapi-mail
websockets
